Holy carp, it's almost there! So I broke down and used PyGame's collision detection, instead of making my own. This is okay for now, but I notice that it's working a little too well. Sometimes two tiles are getting killed instead of just one. I'll check into that soon.
The tile image is just a placeholder for now. Once I get the mechanics working better, I'm going to start designing some levels, so the colors will change. I may just end up using Rects() for this instead of images. We will see. I do want to add an animation to flash when the tile is hit.

The sounds need to be chopped up a bit, I think they are too long. They also seem to lag a second, so IDK what that is about.

But... success!